+++ build.properties
@@ -1,5 +1,5 @@
 mod_version=0.4
 minecraft_version=1.8
-forge_version=11.14.2.1435
-mappings=snapshot_20150606
+forge_version=11.14.3.1487
+mappings=snapshot_20150708
