--- src/mantle/blocks/util/IBlockWithSubtypes.java
@@ -1,8 +1,0 @@
-package mantle.blocks.util;
-
-import net.minecraft.item.ItemStack;
-
-public interface IBlockWithSubtypes
-{
-    String getSubtypeUnlocalizedName(ItemStack stack);
-}
