+++ src/mantle/blocks/util/IBlockWithVariants.java
@@ -1,0 +1,8 @@
+package mantle.blocks.util;
+
+import net.minecraft.item.ItemStack;
+
+public interface IBlockWithVariants
+{
+    String getVariantNameFromStack(ItemStack stack);
+}
